requirements.txt
bottle
pynacl
requests
pyyaml
